Exclude add-on purchases from active that are expiring today
Prior to this change, an add-on purchase was considered still active if
it was expiring today. This doesn't align with how subscriptions are
handled on the CustomersDot side where the end date is excluded from the
current term.

This change aligns the end date check for active add-on purchases to
exclude the end date. Additionally, it now checks the start date as
well. Any add-on purchases without a start date will be considered
active as long as the expires on date is less than today.

EE: true
Changelog: changed