Create missing NPM metadata when package is deprecated
In order to deprecate a package version we need to set attribute deprecated in the package's metadata, so having a metadata entry is a requirement. Currently not all npm packages have associated metadata and this prevents the package from being deprecated. In this MR, when deprecating a package, we create missing metadata entries when they don't exist.

Changelog: fixed