Set the urgency of send_phone_verification_code endpoint to low
This endpoint spends most of its time waiting for the request to
Telesign SMS Verify API endpoint with an average duration of 0.7s. Since
we don't have any control on this endpoint we are very limited on what
we can improve to ensure the current target duration (1s) is met.