List closed child items last
This changes the order for work item child items to account for the
`state` of the work items. We want to order based on state and then
relative_position. This means that open work items will be rendered
first, and closed work items last.

This aligns the behaviour with legacy epic work items.