Retry - add personal_namespace_id column to events
The first attempt was reverted because acquiring a lock on `events` did
not succeed.

Here we add the column in a post-deployment migration for GitLab.com
only, so that we have better control when the migration will be
executed. Once we succeeed there will be subsequent MR to add the column
teh regular way for all other instances.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/164713#note_2090794956

https://gitlab.com/gitlab-org/gitlab/-/issues/462801

Changelog: added