Limit scan execution policy actions to 8
Limits the number of actions that can be added
to a scan execution policy to 8.

We have to merge the CI config for every action.
With a unlimited number of actions, this could lead
to timeout issues.

Currently we support 8 different types of actions so users
will still be able to add an action of every kind.

Changelog: changed
EE: true