Fix bug where car left after branch deletion
When the merge train strategy is 'Add to train when pipeline succeeds'
it's possible to abort the merge after it's been added to the train.
When this happens gitlab currently neglets to destroy the train
and cancel the related pipeline. Instead it only disables the
auto merge.

Changelog: fixed
EE: true