Remove exclusive lock when updating statistics
When updating statistics in the database we wrapped
the operation with `detect_race_on_record` which used
an exclusive lease on the model record to detect
concurrent updates.
We are not taking any actions around concurrent updates
hence we should remove this exclusive lock since
it caused FlushCounterIncrementsWorker to be greatly
affected by Redis load.

Changelog: removed