Add Repository X-Ray service
Adds a service that utilizes
Ai::Context::Dependencies::ConfigFileParser
to scan and parse dependency manager config
files on the repository's default branch.
Upon extracting the files, saves the valid ones
to the `xray_reports` table. Returns a response
with success and error messages.

This service is to be utilized later by a worker
that is triggered on a new commit to the default branch.