Document EventStore disadvantages
We have had two incidents where we had to drop jobs in the end.
As EventStore is built on top of Sidekiq currently, we should make clear
to developers to consider that events are processed asynchrously.

If the event/logic was already processed via a Sidekiq worker, then
there's no change. However if the event/logic is processed synchrously,
and subsequently updated to use EventStore, then the developer should be
aware that the job has a (tiny) chance of not being processed.