Update duo-workflow-executor binary needs to be tar gz
I was originally planning on getting clients to download the exeucutable
file but it turns out this was tricky for our clients to use as the
[docker API](https://docs.docker.com/reference/api/engine/v1.43/#tag/Container/operation/PutContainerArchive)
we are using to copy the file into the container expects an archive to
expand. The `.tar.gz` file has the added benefit that it is about half
the size.