Deprecate non-nullable `target` field on Todo type
Under certain circumstances like the need for SAML-reauthentication,
the target on a Todo can be null.

Unfortunately the target field is not nullable,
so we created a new one target_entity which is nullable.

I still don't understand why `non-nullable` -> `nullable` is a breaking
change on a field because it relaxes a constraint, but here we are.

https://gitlab.com/gitlab-org/gitlab/-/issues/484987

Changelog: changed