Remove SaaS from X509 page
Logically, it's not possible to upload CAs to GitLab.com.
Also, all verification advice on this case deals with local Git commands.
Since customers on GitLab.com likely to expect the `verified` badge to work
(as the mainly visible aspect of this feature), we should not suggest that it can.