Remove legacy E2E jobs
We used a list coming from Snowflake/Tableau,
where we'll fetch pipelines over the last year.

For those use-cases, we need to keep track of
all the CI job names.

set_pipeline_name does not need to have
historical data. It can just work with the
current name of the jobs.