Vulnerability Details: Add feature specs
Adds some feature specs to cover parts of
the vulnerability details page.

The main focus of this change was to add
test-cover for some newly added AI features,
but since there were no existing specs for
this page it should be seen as just a start to
add more coverage in the future.