Add setting for projects to access SPP repository to fetch PEP configs
This change adds a project setting that allows to grant repository
access to SPP project for the linked projects.

This is necessary because user may not have permission to the Security
Policy Project where the linked Pipeline execution policy configuration
is stored. In such case, the user is not able to run pipelines
in projects where the PEP is enabled.

Changelog: added
EE: true