Fix IncidentsList Vue 3 specs
The main problem here is that the child
`PaginatedTableWithSearchAndTabs` component is complex, and apparently
doesn't fully work under Vue 3.

Now, we stub it out, and assert it is passed the expected props.

Fixing `PaginatedTableWithSearchAndTabs` to work under Vue 3 is out of
scope for this change, and will need to be fixed at some point in the
future.