Fix ci_deleted_objects#pick_up_at
This commit fixes a bug where pick_up_at attributes was inheriting
from ci_job_artifacts#expire_at attributes and it was set far away
in the future. We are now defaulting to a much lower value so the
artifacts will get purged instead of staying around forever.

Changelog: fixed