Allow existing users with block seat overages when over seat count
If trying to add an existing user to the group when block seat overages
is enabled, and the number of users exceeds the seat count, then
accept the user.
This may happen if, for instance, a user is currently a developer and
is being added as a maintainer.
Block seat overages is behind a feature flag.