Improved perceived performance on the new merge request dashboard
The database query to get the count for the number of merge requests in a list
can take much longer than the query to get the merge requests.
With this change we request the count separately so that we can load the merge requests
quickly and display them without waiting for the count query to finish.