Detect incompatible use of `Struct.new` in specs
This commit patches `Struct.new` in specs when `keyword_init` is missing
and a Hash passed.

Reason: Since Ruby 3.2 `Struct` supports kwargs by default so
`keyword_init: true` is no longer required.  We are running Ruby 3.2 in
MRs and Ruby 3.1 on master so occasionally master is broken due to this
incompatibility.

This slows down creation of Struct objects in specs by factor 3.