Fix the template for release download redirection
Follow-up for https://gitlab.com/gitlab-org/security/gitlab/-/merge_requests/4407

The MR introduced a warning page when redirection happens
to an external URL (not within the GitLab instance itself).

However, in the HAML template for that page there is a mistake and
the URL in HTML `<a href="">` is not rendered and prints the
raw placeholder instead.

This commit fixes that mistake.