Let embedded videos be full width with max-width 100%
Currently embedded videos are fixed at 400px wide in descriptions and
comments unless a specified width is passed. This is different from the
behavior of images, which default to their natural width with a max of
100%.

This MR removes the default `width="400"` attribute on video elements
and styles them to display at their natural width with a maximum width
of 100% of the parent element, to match the behavior of images.

Changelog: changed