Create add-on purchases when applying an offline license
Prior to this change, add-on purchases were only created for online
cloud licenses during an instance activation or the daily/manual seat
sync.

This change provisions and deprovisions add-on purchases when applying
an offline license as well. The deprovisioning is also already part of
the logic to destroy a license and doesn't require additional changes.

Changelog: added
EE: true