Fix nested CI environment variables for pipeline seed
When passing the environment parameter to the
`scoped_variables_for_pipeline_seed` method, we missed expanding it with
`sort_and_expand_all`. And, this caused not expanding nested variables.

Besides the actual change for the environment,
I also wrote tests cases for k8s variables
to highlight that nested variables are not expanded.