Do not expire direct memberships in subgroups and projects
This commit changes the behaviour when expiring user memberships in
subgroups and projects. Previously, direct memberships were
expired as well. To be consistent with the removal process outlined in
https://docs.gitlab.com/ee/user/group/#remove-a-member-from-the-group
we've changed this behaviour for expiration. Direct memberships are
now left intact.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/471051

Changelog: changed