Show manifest media type for container registry image tag
We added a icon to distinguish Docker manifest lists
and OCI index tags on the container registry with
icon/tooltip text in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/161644

This commit reverts the changes in that MR and hides the size
field for these tags since the size is always 0.
Instead we use a badge with the text 'index'

Additionally, we display the manifest media type for image tags
when the GraphQL api returns them. This is displayed as part
of the collapsible details section

Changelog: changed