Remove unnecessary association validation
Associated members do not need to be validated when
updating a member role. This can cause request
timeouts when many members are associated.
The reason this was added is to ensure the root
namespace of associated members is equal to the root
namespace of the member role. There are other validations
in place which already ensure this.

Changelog: fixed
EE: true