Begin populating the `has_vulnerability_resolution` column
Context
----------------

We need to be able to filter vulnerabilities on whether they have the
"Resolve with Duo" button enabled.

This button is enabled if a finding's `CWE` value is included in this
[hard-coded list][0] of `CWE` values.

We added a new column in [MR 165548][1] to support this feature

This Commit
----------------

This commit starts populating the new column in the vulnerability
ingestion flow.

You can see more detailed discussions in [the epic][2]

[0]:https://gitlab.com/gitlab-org/gitlab/-/blob/1eee1a7b737f56a6f8d4af5c3d864a4838c62560/ee/app/models/vulnerabilities/finding.rb#L21-65
[1]:https://gitlab.com/gitlab-org/gitlab/-/merge_requests/165548
[2]:https://gitlab.com/groups/gitlab-org/-/epics/15036

---

MR: tbd
epic: https://gitlab.com/groups/gitlab-org/-/epics/15036
resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/488394