Avoid database reconnections when host disconnected from load balancer
As described in https://gitlab.com/gitlab-org/gitlab/-/issues/490211,
in Rails 7.0, whenever `ConnectionPool#disconnect!` is called, each
connection is acquired by the thread and verified with a SQL `;`
query. If the verification fails, then Rails will attempt a reconnect
for all the connections in the pool. This reconnection can cause
unnecessary database connection saturation and result in a flood of
SET statements on a PostgreSQL host when many threads attempt the same
thing.

Rails 7.1 has fixed this in https://github.com/rails/rails/pull/44576,
but until we upgrade this patch disables this verification step.

This commit introduces a `load_balancing_disconnect_without_verify`
feature flag to enable this change.

Changelog: fixed