Placeholder to avoid signup restriction validation
This validation can block placeholder and import users to be created on
systems that have sign up restrictions.

However, placeholder users and import users are essential to be created
by the system during imports.

Other services will set a `created_by_id` for the user record which
avoids the validation, as placeholder and import users are created by
the system there is no `created_by_id` we could set.

The sign up restrictions do not need to apply to these users because
they're not accounts that can be signed in by users due to being
internal user types.

https://gitlab.com/gitlab-org/gitlab/-/issues/483639