Fix truncated messages for commits
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/357459

**Problem**

Code responsible for displaying a list of last commits for each path
truncates the commit message to 1Kb. The same value is used to generate
a markdown cache for commit description. It create a cache with a
limited description.

**Solution**

Trigger the cache creation before the description is truncated to be
stored in Redis cache.

Changelog: fixed