E2E: Log Interceptor Api Errors at warn level
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/81613 introduced
the ability to capture api error metadata. However sometimes this data
is not available and we end up logging an a message at the error level.

These messages being logged at the error level make scrolling through
the logs of our tests a little misleading as you can think there is an
error, but the test is actually working. We should consider reserving
the use of error level logging for failures that will impact the test
results, and therefore it seems more appropriate to log these messages
at the WARN level.