Enable external status checks mutation for Projects::AllBranchesRule
We accidentally disabled the graphql mutations for BranchRules
ExternalStatusChecks for Projects::AllBranchesRule records. This is
incorrect, we do allow external status checks to be configured for `All
branches`.

This change updates the service classes to allow external status checks
for all branches. It also updates the mutation specs to ensure this is
possible through the GraphQL API.

Changelog: fixed
EE: true