Check for banned user normalized email reuse as part of model validation
Instead of triggering the check for banned user normalized email reuse
in a before_action hook, trigger it as part of the User model
validation step and only when there are no other email validation errors
already present. This prevents the additional DB query for the check
which results to a lower priority action prompt for the user compared to
the earlier validation errors (e.g. email is not unique).