Add workflowStatus to DuoWorkflowEvent graphql subscription
Duo Workflow is a new feature where you can start running an AI workflow
from VS Code (and other places in the future). The UI for this streams
events about what is happening from a GraphQL susbcription.

Related to
https://gitlab.com/gitlab-org/editor-extensions/gitlab-lsp/-/issues/401
we want to be able to update the workflow status as we stream events
while the workflow is running. We already subscribe to these events so
this is a good place to put it.