Add API endpoint to publish CI Catalog Resource Version
Today, when publishing a new version to the CI Catalog, the metadata of
components are fetched from the repository and extracted from the files.

The goal is to move this responsibility to `release-cli` and accept
the required metadata from the client. To this end, a new API endpoint
will accept `components_data` and pass it to
`Ci::Catalog::Resources::ReleaseService`.

In this MR, we are implementing the new API endpoint
`POST /projects/:id/catalog/publish` to publish a CI Catalog Resource
Version.

Changelog: added