Todos: Add namespace and project target types to GraphQL
If you have a member access request for a project, the target type will
be a project or a namespace. The GraphQL API didn't know about this,
leading to errors.

We also add a spec that makes sure that all known types for the
TodoFinder are exposed in the GraphQL enum.

Changelog: fixed