Backfill `request_accepted_at` column to `created_at`
Previously, the system was using the request timestamp instead of an
actual grant timestamp when displaying "given access" date to the user.

This commit introduces a backfill migration of `request_accepted_at`
column to the `created_at` in the members table. The change allows us to
display the correct timing for when access was actually granted to the
user, improving the accuracy of the user interface and avoiding
confusion about access timing.

This commit is a part of an ongoing work. The related MRs are:
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/156421
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/152584

Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/24721
Changelog: added