doc: Use Mermaid for Simplified Component Overview
The Simplified Component Overview was previously made using Miro.

There's no reason to have this external dependency, and it was also
unclear who had access to update the Miro board. Instead, convert the
diagram to Mermaid so it's in-line with the detailed component diagram
in the next section.

Also fixes some errors concerning Gitaly:

- Gitaly does not communicate with Redis at all, so remove that link.
- Gitaly communicates with Workhorse when it calls Rails APIs. Add that
  link.