Backfill for `vulnerability_reads.has_vulnerability_resolution`
Context
--------------

We need to be able to filter vulnerabilities on whether they have the
"Resolve with Duo" button enabled.

This button is enabled if a finding's `CWE` value is included in this
[hard-coded list][0] of `CWE` values.

In previous MRs we:

- Added a new column to support this filter [^1]
- Began populating this column for new rows [^2]
- Added a scope to filter on thid column [^3]

This commit
--------------

This commit adds the [batched background migration][1] to backfill the
existing rows.

**Note**

In the [model code][2], we define `cwe_value` like this:

```ruby
    def cwe_value
      identifiers.find(&:cwe?)&.name
    end
```

`find` will stop at the first `vulnerability_identifier` that has an
`external_id == CWE`

So, given the following scenario:

- We have a `vulnerability_finding` that has 2 `identifiers`
- both `identifiers` are `cwe` identifiers
- the first identifier is NOT in `HIGH_CONFIDENCE_AI_RESOLUTION_CWES`,
  but the second is

Then the application code would return `false` for
`ai_resolution_enabled?`, whereas this backfill will mark it as `true`

Our current business logic for `SAST` rules does not actually allow
multiple CWE identifiers, so this is not a real scenario
in production [^4]

[0]:https://gitlab.com/gitlab-org/gitlab/-/blob/05b60865a750e1244eea1910ffd4d02e9c462410/ee/app/models/vulnerabilities/finding.rb#L25-70
[1]:https://docs.gitlab.com/ee/development/database/batched_background_migrations.html
[2]:https://gitlab.com/gitlab-org/gitlab/-/blob/7d3bfaf874e409284d3d465b3ff91f383472585a/ee/app/models/vulnerabilities/finding.rb#L429

[^1]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/165548
[^2]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/165952
[^3]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/166312
[^4]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/166110#note_2120772688

---

Changelog: added
EE: true
Resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/486530
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/166110
Epic: https://gitlab.com/groups/gitlab-org/-/epics/15036