WorkItem: Add StartAndDueDate widget EE extension
This is the first step towards the unification of work item
`RolledupDate` widget and `StartAndDueDate` widget. In here we're
creating an `EE` extension for the `StartAndDueDate` widget on the
GraphQL layer so the ~frontend can use the same Type to manage fixed
dates and rolled up dates.

For this, the `StartAndDueDate` now contains a `canRolleup` to indicate
if the WorkItem can use rolled up dates or only fixed dates. When
`canRolleup` is true, we can render a radiobox to let the users choose
between fixed dates or rolled up dates.

Examples of the `StartAndDueDateType` payload:

- on `FOSS`:

```json
        {
          "type": "START_AND_DUE_DATE",
          "canRollup": false,
          "startDate": null,
          "dueDate": null
        }
```

- on `EE`:

```json
        {
          "type": "START_AND_DUE_DATE",
          "canRollup": true,
          "isFixed": false,
          "startDate": "2024-09-22",
          "startDateSourcingMilestone": null,
          "startDateSourcingWorkItem": {
            "id": "gid://gitlab/WorkItem/653"
          },
          "dueDate": "2024-09-15",
          "dueDateSourcingMilestone": null,
          "dueDateSourcingWorkItem": {
            "id": "gid://gitlab/WorkItem/653"
          }
        }
```

Related to: https://gitlab.com/gitlab-org/gitlab/-/work_items/486526
Changelog: added
EE: true