Add emails notification for RewriteHistory worker
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/483085

**Problem**

Synchrounous rewriting of repository history is too slow. We are going
to make it asynchronous instead. But it's currently missing a
notification channel for users to receive updates about the processing
status.

**Solution**

Define email templates that is going to be called from RewriteHistory
Sidekiq job.

Changelog: other