Fix Vue Translation Mixin specs
The jest specs for the Vue translation mixins were using the runtime
compiler of Vue, because they were defining components within the specs
which used the `template:` interface.

When moving the templates to an actual componetn which goes through our
usual compiler chain, the specs were fixed automatically.

We could make this even nicer, but who cares?