Fix draft not clearing after submitting description
When editing the description of a work item, a draft is saved on
every key press. When editing is cancelled by the user, the draft
is deleted. However, when editing is saved, the draft is not deleted.
This means, in some cases, this old draft is used when the user starts
editing. The solution is to clear the draft when editing is saved.

Changelog: fixed