Fix WorkspaceVariables Vue 3 template compilation
The compilation error was:

    SyntaxError: v-else/v-else-if has no adjacent v-if or v-else-if.

The problem was that we had a v-if/v-else on different slots, i.e.:

    <template v-if="..." #foo>...</template>
    <template v-else="..." #bar>...</template>

In Vue 3, this doesn't compile, presumably because the node representing
the v-if doesn't exist by the time the v-else is encountered. In Vue 2,
this pattern is fine.

If the same slot name is used, it compiles just fine:

    <template v-if="..." #foo>...</template>
    <template v-else="..." #foo>...</template>

I suspect this is an accidental implementation detail.