Ignore warning caused by vue-demi
It turns out we have a transitive dependency on `vue-demi` now via
`pinia`.

The `vue-demi` package has a postinstall script which checks the
installed `vue` package version, and copies files around accordingly.

For us, `vue` is always 2.x, and we rely on the `VUE_VERSION`
environment variable to change Jest's behaviour and dynamically map
`vue` to `@vue/compat`. This means `vue-demi` needs to be explicitly
told which Vue version we're using.

This is quite messy, but for now, this only matters for the
jest-build-cache jobs. As `pinia` gets more deeply integrated into the
codebase, or we rely on `vue-demi` in other ways, we'll probably want to
create a script to for switching between Vue 2 and 3 modes (or similar).

See https://gitlab.com/groups/gitlab-org/-/epics/15340.

For now, this will do.