Disallow predictive Vue 3 Jest failures
This should help reduce regressions in Vue 3 Jest compatibility.
These predictive jobs only run specs which are affected by the changes
in the given merge request, minus the specs that are known to fail under
Vue 3.

If somehow a predictive job fails in a way that's unrelated to the MR,
the MR isn't blocked. This is because once it reaches ~pipeline::tier-3,
the full `jest.* vue3 mr` jobs will run, which *are* allowed to fail.