Fix timeline style in dark mode
In dark mode, the vertical line representing the timeline in an
activity feed was a different color than the background in the places
where it should be hidden. This is likely a result of incomplete
refactoring when the default background color was applied. This is
easily fixed by removing a css rule.

Changelog: fixed