Order labels in system note by label name
Properly sort the labels by label name. Right now we sort by reference
HTML which leads to scoped labels being sorted by their hex color.

This also fixes flaky specs asserting on the ID references because IDs
in tests are not fixed and we are sorting the references as strings so
~10 gets sorted ahead of ~9.