Expose `has_vulnerability_resolution` via GraphQL
Context
-------------

We need to be able to filter vulnerabilities on whether they have the
"Resolve with Duo" button enabled.

This button is enabled if a finding's `CWE` value is included in this
[hard-coded list][0] of `CWE` values.

In previous MRs we:

-   Added a new column to support this filter [^1]
-   Began populating this column for new rows [^2]

This commit
-----------------

The database work isn't done yet as we still need to backfill the
existing rows[^3] and add database indexes[^4].

However, we can expose graphql fields with FFs according to the docs:

-   [*Backend GraphQL API guide: Feature Flags*][1]
-   [*Backend GraphQL API guide: Mark schema items as Alpha*][2]

This will unblock the frontend team from building and testing their
code, as well as help us solidify the API contract.

[0]:https://gitlab.com/gitlab-org/gitlab/-/blob/1eee1a7b737f56a6f8d4af5c3d864a4838c62560/ee/app/models/vulnerabilities/finding.rb#L21-65
[1]:https://docs.gitlab.com/ee/development/api_graphql_styleguide.html#descriptions-for-feature-flagged-items
[2]:https://docs.gitlab.com/ee/development/api_graphql_styleguide.html#mark-schema-items-as-alpha

[^1]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/165548
[^2]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/165952
[^3]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/166110
[^4]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/165548#note_2097036644

---

Changelog: added
EE: true
Resolves: https://gitlab.com/gitlab-org/gitlab/-/issues/490323 <br>
Epic: https://gitlab.com/groups/gitlab-org/-/epics/15036 <br>
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/166312