Skip new branch name validation for some keypresses
Whenever the user releases any key while the new branch name field is
focused on an issue, the branch name is re-validated, even if no change
was made to the value.

This MR excludes keyups for certain key combinations so that copying the
branch name, using the arrow keys and hitting escape do not trigger
unnecessary validations.

Changelog: fixed