Apply force push settings consistently
We recently implemented group level protected branches. The feature is
still behind a disabled feature flag. During testing it was noted that
force push settings are applied differently to access levels.

When group-level access levels exist, we check if any of the
group-or-project-level access levels allow a user to do said action.
Before this change, when a group-level protected branch existed we would
completely ignore any project level protected branches.

After this change we check against all matching protected branches when
evaluating if a force push is allowed or not.