Setup PostReceive worker for sticky data behind flag
We have introduced a new feature flag to rollout the sticky data
consistency for PostReceive worker.

This worker handles a lot of traffic as it is triggered after each push
to GitLab.

For safety we are enabling stick consistency behind a feature flag named
load_balanacing_for_sticky_post_receive_worker. This is described in
this documentation for workers:
https://docs.gitlab.com/ee/development/sidekiq/worker_attributes.html#feature_flag-property

Closes https://gitlab.com/gitlab-org/gitlab/-/issues/402254