Use component_names instead of component_ids
We recently switched to returning component names instead
of ids in the filtered search as the backend returns
de-duplicated results with the names. To match the same
behaviour with the dependencies results, we have to send
the component names instead of ids. This commit implements
the necessary changes for this fix.