Show specific OmniAuth buttons in signup page
If the omniauth option `allow_single_sign_on` is `false` hide the
OmniAuth buttons in signup page. That's because in this scenario an
account must be created previously.

If `allow_single_sign_on` is an array, show only those providers in the
signup page.

See
https://docs.gitlab.com/ee/integration/omniauth.html#configure-common-settings

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/415478

Note: there is a setting that allows to signup, that's different from
`allow_single_sign_on`.

Changelog: changed