Only show custom emoji from current group in frequently used emoji
Currently, the user's frequently used emojis group shows their 9 most
recently used emoji, regardless of whether they are custom emoji in the
current gorup or not. This means that custom emojis that are not from
the current group will show as a "?" placeholder in the frequently used
category.

This MR filters those emojis out of the Frequently Used category.

Changelog: fixed