Add support for multiples files for Python Pip config file class
Adds support to find and parse multiple files for the
PythonPip config class. This is because Pip may have
more than one 'requirements' file.

This means that the ConfigFileParser may produce an
array of objects where more than one object may have
the same lang value.

In the next MR, we will update the Repository X-Ray
ScanDependencies service so that it combines the
payloads of config file objects belonging to the
same lang.