Add upload authorize struct support within the senddependency command
Add support in the workhorse senddependency action to include the
upload authorization response.

This way, the upload handler doesn't need to get the upload
authorization from the rails backend, it's already there.

This saves one request to the rails backend and shorten the overall
upload processing time.

Apply this shortcut in the maven virtual registry feature.