Fixes an invalid refType being sent by frontend
Sometimes a `refType` can be sent to the GraphQL API with an invalid type
which in turn causes a GraphQL error.
This changes that by making sure we only ever send to the API what the API will expect.
If the value from the URL (or whatever else) is invalud then we just send `null`.

https://gitlab.com/gitlab-org/gitlab/-/issues/480353