Remove project_id FK constraint on terraform module metadata
The project_id is the sharding key in the packages_terraform_module_metadata table, so it cannot be nullified even if the parent project got deleted. Thus, in this MR, we remove the FK constraint so that the project can be deleted if if it has associated packages_terraform_module_metadata records.

Changelog: fixed