Add a hasAiResolution filter to vulnerabilitySeveritiesCountResolver
We need to be able to filter vulnerabilities on whether they have the
"Resolve with Duo" button enabled.

This button is enabled if a finding's `CWE` value is included in this
[hard-coded list][0] of `CWE` values.

In a previous MRs we exposed this filter in the
`VulnerabilitiesResolver`[^1]

This commit updates the `vulnerabilitySeveritiesCountResolver` to make
use of the same filter.

[^1]: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/166312

---

Changelog: changed
EE: true
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/167433
epic: https://gitlab.com/groups/gitlab-org/-/epics/15036
Related to: https://gitlab.com/gitlab-org/gitlab/-/issues/481529