Add a GraphQL mutation for updating `OrganizationUser`
This change introduces a new GraphQL mutation for updating
`OrganizationUser`. The mutation is restricted for use by
organization owners only. The access level can be set to either
`OWNER` or `DEFAULT`. To prevent organizations from being left
without an owner, the role of the last owner cannot be changed.

Changelog: added