Improve scope of Feature-Enabled-By-Namespace-Ids header
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/162183 we
added the X-Gitlab-Feature-Enabled-By-Namespace-Ids header and populated
it with the namespace ids of the Add-On purchases assigned to the user.
But the authorization logic is more involved than that, and before each
service's cut-off date, an AI feature can also be enabled based on the
plan in the namespaces the user belongs to. This change takes advantage
of the unified auth logic introduced in
https://gitlab.com/gitlab-org/gitlab/-/issues/473087, and adapts the
`User.allowed_to_use?` method to have it yield the namespace ids that
allow access to the AI feature, if any.

Note: at the moment, the full Duo Chat auth logic has not been migrated
to the new unified auth code, so the header will still be inaccurate for
that featur. See https://gitlab.com/gitlab-org/gitlab/-/issues/497638.